requirements.txt
Flask==3.0.3
matplotlib==3.9.2
pandas==2.2.3
seaborn==0.13.2
tensorflow-intel==2.17.0
waitress==3.0.0
scikit-learn==1.5.2
